requirements.txt
Django
djangorestframework
Pillow
drf-yasg
psycopg2-binary
python-decouple
django-filter
